requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
websocket-client>=1.7.0
statsmodels>=0.14.0
scipy>=1.10.0
watchdog>=3.0.0
